docs: generated comments to docs.rs
