+++ src/utils.rs
@@ -1,4 +1,6 @@
+/*!
+    Module containing utility implementations for [crate::EndpointUtils] and API response handling.
+*/
+
 pub mod endpoint;
 pub mod response_handler;
-
-pub use endpoint::EndpointUtils;
